Implement System Properties
This implements the system properties needed for AppFit. This includes all of the device info, operating system, and versioning info the SDK can obtain quickly and smoothly. This also support the IP and GeoLocation info that is unifhsed. This implements all of the models and keys used in the properties.